+++ demo/api/src/db/fixtures/generators/unsplash-image-file-fixture.service.ts
@@ -1,9 +1,12 @@
-import { download, FileInterface, FilesService } from "@comet/cms-api";
+import { FileInterface, FilesService, FileUploadService } from "@comet/cms-api";
+import { Injectable } from "@nestjs/common";
 import { DamScope } from "@src/dam/dto/dam-scope";
 import faker from "faker";
 
-export class UnsplashImageFileFixture {
-    constructor(private filesService: FilesService) {}
+@Injectable()
+export class UnsplashImageFileFixtureService {
+    constructor(private readonly filesService: FilesService, private readonly fileUploadService: FileUploadService) {}
+
     async generateImage(scope: DamScope): Promise<FileInterface> {
         const width = faker.datatype.number({
             min: 1000,
@@ -16,7 +19,7 @@ export class UnsplashImageFileFixture {
 
         const imageUrl = `https://source.unsplash.com/all/${width}x${height}`;
         console.log(`Downloading ${imageUrl}.`);
-        const downloadedImage = await download(imageUrl);
+        const downloadedImage = await this.fileUploadService.createFileUploadInputFromUrl(imageUrl);
         console.log(`Downloading ${imageUrl} done.`);
         console.log(`Uploading ${downloadedImage.originalname}.`);
         const file = await this.filesService.upload(downloadedImage, { scope });