+++ demo/site/src/util/createGraphQLClient.ts
@@ -1,12 +1,18 @@
-import { PreviewData } from "@comet/cms-site";
 import { GraphQLClient } from "graphql-request";
 
-export default function createGraphQLClient(previewData?: PreviewData): GraphQLClient {
-    const { includeInvisibleBlocks, includeInvisiblePages, previewDamUrls } = {
-        includeInvisiblePages: !!previewData,
-        includeInvisibleBlocks: previewData && previewData.includeInvisible,
-        previewDamUrls: !!previewData,
-    };
+export type GraphQLClientOptions = {
+    includeInvisiblePages: boolean;
+    includeInvisibleBlocks: boolean;
+    previewDamUrls: boolean;
+};
+
+const defaultOptions: GraphQLClientOptions = {
+    includeInvisiblePages: false,
+    includeInvisibleBlocks: false,
+    previewDamUrls: false,
+};
+export default function createGraphQLClient(options: Partial<GraphQLClientOptions> = {}): GraphQLClient {
+    const { includeInvisibleBlocks, includeInvisiblePages, previewDamUrls } = { ...defaultOptions, ...options };
 
     const headers: Record<string, string> = {};