+++ packages/admin/cms-admin/src/dam/DataGrid/fileUpload/useDamFileUpload.tsx
@@ -34,7 +34,7 @@ interface FileWithPath extends File {
     path?: string;
 }
 
-interface FileWithFolderPath extends FileWithPath {
+export interface FileWithFolderPath extends FileWithPath {
     folderPath?: string;
 }
 
@@ -55,7 +55,10 @@ interface UploadFilesOptions {
 }
 
 export interface FileUploadApi {
-    uploadFiles: ({ acceptedFiles, fileRejections }: Files, { folderId, importSource }: UploadFilesOptions) => Promise<void>;
+    uploadFiles: (
+        { acceptedFiles, fileRejections }: Files,
+        { folderId, importSource }: UploadFilesOptions,
+    ) => Promise<{ hasError: boolean; rejectedFiles: RejectedFile[]; uploadedItems: NewlyUploadedItem[] }>;
     validationErrors?: FileUploadValidationError[];
     maxFileSizeInBytes: number;
     dialogs: React.ReactNode;
@@ -70,6 +73,10 @@ export interface FileUploadApi {
 export interface FileUploadValidationError {
     file: Pick<FileWithFolderPath, "name" | "path">;
     message: React.ReactNode;
+}
+
+interface RejectedFile {
+    file: File;
 }
 
 const addFolderPathToFiles = async (acceptedFiles: FileWithPath[]): Promise<FileWithFolderPath[]> => {
@@ -358,17 +365,22 @@ export const useDamFileUpload = (options: UploadDamFileOptions): FileUploadApi =
         [manualDuplicatedFilenamesHandler],
     );
 
-    const uploadFiles = async ({ acceptedFiles, fileRejections }: Files, { folderId, importSource }: UploadFilesOptions): Promise<void> => {
+    const uploadFiles = async (
+        { acceptedFiles, fileRejections }: Files,
+        { folderId, importSource }: UploadFilesOptions,
+    ): Promise<{ hasError: boolean; rejectedFiles: RejectedFile[]; uploadedItems: NewlyUploadedItem[] }> => {
         setProgressDialogOpen(true);
         setValidationErrors(undefined);
 
         const uploadedFolders: Array<NewlyUploadedItem & { type: "folder" }> = [];
         const uploadedFiles: Array<NewlyUploadedItem & { type: "file" }> = [];
+        const rejectedFiles: Array<RejectedFile> = [];
 
         let errorOccurred = false;
         if (fileRejections.length > 0) {
             errorOccurred = true;
             generateValidationErrorsForRejectedFiles(fileRejections);
+            rejectedFiles.push(...fileRejections.map(({ errors, ...rejection }) => rejection));
         }
 
         if (acceptedFiles.length > 0) {
@@ -415,7 +427,7 @@ export const useDamFileUpload = (options: UploadDamFileOptions): FileUploadApi =
                         },
                     );
 
-                    uploadedFiles.push({ id: response.data.id, parentId: targetFolderId, type: "file" });
+                    uploadedFiles.push({ id: response.data.id, parentId: targetFolderId, type: "file", file });
                 } catch (err) {
                     errorOccurred = true;
                     const typedErr = err as AxiosError<{ error: string; message: string; statusCode: number }>;
@@ -442,6 +454,8 @@ export const useDamFileUpload = (options: UploadDamFileOptions): FileUploadApi =
                     } else {
                         addValidationError(file, <UnknownError />);
                     }
+
+                    rejectedFiles.push({ file });
                 }
             }
 
@@ -457,6 +471,8 @@ export const useDamFileUpload = (options: UploadDamFileOptions): FileUploadApi =
         setUploadedSizes({});
 
         addNewlyUploadedItems([...uploadedFolders, ...uploadedFiles]);
+
+        return { hasError: errorOccurred, rejectedFiles: rejectedFiles, uploadedItems: [...uploadedFolders, ...uploadedFiles] };
     };
 
     return {