+++ packages/admin/cms-admin/src/generator/future/generator.ts
@@ -35,7 +35,7 @@ export type GeneratorConfig = FormConfig<any> | TabsConfig;
 
 export type GeneratorReturn = { code: string; gqlQueries: Record<string, string> };
 
-async function main() {
+export async function runFutureGenerate() {
     const schema = await loadSchema("./schema.gql", {
         loaders: [new GraphQLFileLoader()],
     });
@@ -75,4 +75,3 @@ async function main() {
         }
     }
 }
-main();