+++ packages/admin/cms-admin/src/generator/index.ts
@@ -1,1 +1,0 @@
-export { FormConfig } from "./future/generator";