+++ packages/admin/cms-admin/src/sitesConfig/useSitesConfig.ts
@@ -2,7 +2,7 @@ import * as React from "react";
 
 import { SiteConfigApi, SiteConfigContext } from "./SitesConfigContext";
 
-export function useSitesConfig(): SiteConfigApi {
+export function useSitesConfig<Configs = unknown>(): SiteConfigApi<Configs> {
     const context = React.useContext(SiteConfigContext);
 
     if (!context) {
@@ -11,5 +11,5 @@ export function useSitesConfig(): SiteConfigApi {
         );
     }
 
-    return context;
+    return context as SiteConfigApi<Configs>;
 }