+++ packages/api/cms-api/src/dam/files/files.service.ts
@@ -31,7 +31,7 @@ import { DamFileListPositionArgs, FileArgsInterface } from "./dto/file.args";
 import { UploadFileBodyInterface } from "./dto/file.body";
 import { CreateFileInput, ImageFileInput, UpdateFileInput } from "./dto/file.input";
 import { FileParams } from "./dto/file.params";
-import { FileUploadInterface } from "./dto/file-upload.interface";
+import { FileUploadInput } from "./dto/file-upload.input";
 import { FILE_TABLE_NAME, FileInterface } from "./entities/file.entity";
 import { DamFileImage } from "./entities/file-image.entity";
 import { FolderInterface } from "./entities/folder.entity";
@@ -302,7 +302,7 @@ export class FilesService {
     }
 
     async upload(
-        file: FileUploadInterface,
+        file: FileUploadInput,
         { folderId, scope, ...assignData }: Omit<UploadFileBodyInterface, "scope"> & { scope?: DamScopeInterface },
     ): Promise<FileInterface> {
         let result: FileInterface | undefined = undefined;
@@ -462,7 +462,7 @@ export class FilesService {
         if (!inboxFolder) {
             throw new Error("Specified inbox folder doesn't exist.");
         }
-        if (inboxFolder.scope && !this.accessControlService.isAllowedContentScope(user, inboxFolder.scope)) {
+        if (inboxFolder.scope && !this.accessControlService.isAllowed(user, "dam", inboxFolder.scope)) {
             throw new Error("User can't access the target scope");
         }
 
@@ -488,7 +488,7 @@ export class FilesService {
 
         const fileScopes = getUniqueFileScopes(files);
         const canAccessFileScopes = fileScopes.every((scope) => {
-            return this.accessControlService.isAllowedContentScope(user, scope);
+            return this.accessControlService.isAllowed(user, "dam", scope);
         });
         if (!canAccessFileScopes) {
             throw new Error(`User can't access the scope of one or more files`);